test(parser): split test_parser.py into a per-language package
Break the 1418-line tests/unit/ingestion/test_parser.py monolith into a
tests/unit/ingestion/parser/ package: the module-scoped parser fixture
moves to conftest.py, _make_file_info to _helpers.py, and each language's
inline source constant + Test class into its own module (python,
typescript, go, rust, java, cpp, kotlin, ruby, csharp, swift, scala, php,
luau) plus test_misc.py for the unsupported-language and LanguageConfigs
cases. Test bodies moved verbatim; same 110 tests collected and passing.
