+++ tests/unit/ingestion/parser/test_cpp.py
@@ -1,0 +1,75 @@
+"""Unit tests for the unified ASTParser.
+
+Tests parse inline byte strings so no filesystem I/O is needed.
+Covers Python, TypeScript, Go, Rust, Java, C++ — one test class per language.
+"""
+
+from __future__ import annotations
+
+from repowise.core.ingestion.parser import ASTParser
+from tests.unit.ingestion.parser._helpers import _make_file_info
+
+CPP_SOURCE = b"""#include "calculator.hpp"
+#include <stdexcept>
+#include <string>
+
+namespace sample {
+
+double Calculator::add(double x, double y) {
+    return x + y;
+}
+
+double Calculator::divide(double x, double y) {
+    if (y == 0.0) {
+        throw std::invalid_argument("Division by zero");
+    }
+    return x / y;
+}
+
+}  // namespace sample
+"""
+
+
+CPP_HEADER_SOURCE = b"""#pragma once
+
+#include <vector>
+#include "models.hpp"
+
+namespace sample {
+
+class Calculator {
+public:
+    double add(double x, double y);
+    double subtract(double x, double y);
+    double divide(double x, double y);
+
+private:
+    std::vector<int> history_;
+};
+
+}  // namespace sample
+"""
+
+
+class TestCppParser:
+    def test_finds_class_in_header(self, parser: ASTParser) -> None:
+        fi = _make_file_info("cpp_pkg/calculator.hpp", "cpp")
+        result = parser.parse_file(fi, CPP_HEADER_SOURCE)
+        classes = [s for s in result.symbols if s.kind == "class"]
+        assert any(s.name == "Calculator" for s in classes)
+
+    def test_finds_functions_in_source(self, parser: ASTParser) -> None:
+        fi = _make_file_info("cpp_pkg/calculator.cpp", "cpp")
+        result = parser.parse_file(fi, CPP_SOURCE)
+        # ``Calculator::add`` style qualified definitions now resolve to
+        # ``kind=method`` with ``parent_name=Calculator``; free functions
+        # stay ``kind=function``. Either way we expect symbols to land.
+        callable_symbols = [s for s in result.symbols if s.kind in ("function", "method")]
+        assert len(callable_symbols) >= 1
+
+    def test_parses_includes(self, parser: ASTParser) -> None:
+        fi = _make_file_info("cpp_pkg/calculator.cpp", "cpp")
+        result = parser.parse_file(fi, CPP_SOURCE)
+        assert len(result.imports) >= 2
+        module_paths = [i.module_path for i in result.imports]
+        assert any("calculator.hpp" in p or "stdexcept" in p for p in module_paths)
